backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy==2.0.31
pymysql==1.1.1
python-dotenv==1.0.1
cryptography==43.0.1

# auth / security
python-jose==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2   # compatible version

# pydantic (v2)
pydantic==2.9.2
pydantic-settings==2.6.1   # settings management (with python-dotenv)

# etc
typing-extensions>=4.12.2

# http requests
requests==2.32.3